import: Convert arrow function to regular function
This fixes a bug introduced in commit
b711ff6acf976a3f8808325411cf244d55c5a0e7. Some time between when that
commit was originally written and when it was merged a round of
linting had converted the function from a regular function to an arrow
function because `this` was never in the body of the function. When I
rebased the commit, which introduced `this` to the body, I didn't
catch the error.