Let Express log HTTP handler errors
Express already comes with a built-in error handler that is attached
as the last middleware. Any handler or middleware can pass an Error
object (with an optional `.status` or `.statusCode` property set to a
4xx or 5xx code) to `next()` and Express will both log a detailed
error message and display a useful error message to the user. For more
details see "The default error handler" at:
https://expressjs.com/en/guide/error-handling.html