ace2_inner.js: Do fn+delete customly
'fn+delete' (or 'delete' in PCs) behaviour is inconsistent across browsers. In Chrome,
it might remove the class of the text in the line below when using
fn+delete to join lines. In this PR, we solve this issue by
doing 'fn+delete' customly, as we do 'delete' (or 'backspace' in PCs) customly.
Fixes: #5172.