express: Move `preAuthorize` hook after `express-session`
The `ep_openid_connect` plugin needs access to session state before
authorization checks are made (to securely redirect the user back to
the start page when authentication completes). Now that the
`expressPreSession` hook exists, the rationale for moving
`preAuthorize` before the `express-session` middleware is gone.

This change undoes the following commits:
  * bf35dcfc5096a77b19bc2b50261b34c98fce7011
  * 0b1ec20c5c061c340db2e7feac2229d79d960296
  * 30544b564eb7658c6830ba38a31af856fd6509fc