fix #6495 failing Docker build when checked out as git submodule
When checked out as a git submodule, `.git` is not a folder containing
`refs` and `HEAD`, but a file pointing to the parent repositories' git
metadata.

This commit fixes the failing Docker build by making the file references
copied from the `.git` folder glob patterns, which don't fail when no
source file can be found.

I could not resist to fix some linter errors in the `Dockerfile` along
the way.