Do not respond to right alt (Alt Gr) keyboard events.
On some keyboard layouts (eg. pl) the right alt key (a.k.a. Alt Gr) is
used in combination with letters to type special characters, eg. AltGr-C
makes a 'ć' on a pl keyboard layout.

At least on Firefox on macOS it seems like events that would otherwise
result in a special character still get passed to keydown and friends.
This means that if etherpad handles alt-c, it might swallow someone's
intent to type in 'ć' and instead open up the chat window.

This could arguably be a bug in Firefox or macOS, as the same doesn't
happen in Firefox on Linux/X11. But before I bring the discussion to a
wider audience I'd like to at least make the issue go away at etherpad
level.